For the start of element and end of element events, it is possible that the default handler does not get called even when no element handlers are called, because storeAtts() may trigger the startNamespaceDeclHandler where the application has a chance to clear the element handlers.
This patch fixes this with one issue still open, which applies to empty elements only:
When the endElement handler is called, but not the startElementHandler,
then the default handler is not called, which causes attributes to not
be reported at all. The other alternative would be to call the default handler,
but then the element would be reported twice.